requirements.txt
numpy
matplotlib
imageio
scipy
pypng
ruamel.yaml
cython
pyopengl
triangle
glumpy
opencv-python
pillow
